+++ src/EventManager.php
@@ -2,6 +2,8 @@
 
 namespace Pulsar;
 
+use Pulsar\Event\AbstractEvent;
+use Pulsar\Exception\ListenerException;
 use Symfony\Component\EventDispatcher\EventDispatcher;
 
 /**
@@ -32,5 +34,47 @@ class EventManager
         if (isset(self::$dispatchers[$class])) {
             unset(self::$dispatchers[$class]);
         }
+    }
+
+    /**
+     * Subscribes to a listener to an event.
+     *
+     * @param string $event    event name
+     * @param int    $priority optional priority, higher #s get called first
+     */
+    public static function listen(string $class, string $event, callable $listener, int $priority = 0): void
+    {
+        self::getDispatcher($class)->addListener($event, $listener, $priority);
+    }
+
+    /**
+     * Dispatches the given event and checks if it was successful.
+     * If it fails then any active transaction will be rolled back
+     * on the model.
+     *
+     * @return bool true if the events were successfully propagated
+     */
+    public static function dispatch(Model $model, AbstractEvent $event, bool $usesTransactions): bool
+    {
+        // Model events can fail whenever $event->stopPropagation() is called
+        // or when a specific exception type is thrown by the listener.
+        try {
+            self::getDispatcher(get_class($model))->dispatch($event, $event::NAME);
+        } catch (ListenerException $e) {
+            // Listener exceptions provide the error message to be stored on the model
+            $model->getErrors()->add($e->getMessage(), $e->getContext());
+            $event->stopPropagation();
+        }
+
+        if (!$event->isPropagationStopped()) {
+            return true;
+        }
+
+        // when listeners fail roll back any database transaction
+        if ($usesTransactions) {
+            $model::getDriver()->rollBackTransaction($model->getConnection());
+        }
+
+        return false;
     }
 }